Felhantering av sidebarsen

+++ Wordless/Wordless/Controllers/AuthorPageController.cs
@@ -74,7 +74,7 @@ namespace Wordless.Controllers
             SetListsForViews();
             if (ModelState.IsValid)
             {
-                var userId = (int)Session["currentUserId"];
+                var userId = (Int64)Session["currentUserId"];
                 var user = db.Users.Find(userId);
                 var newFile = new File
                 {
